requirements.txt
pandas
scikit-learn
langchain
langgraph
pypdf
python-docx
reportlab
pdfplumber
pymupdf
pytesseract
pillow
requests
beautifulsoup4
google-generativeai
groq
joblib
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pymongo
dnspython